remove ledger splash component + move ledger utils
